ported display component to src folder. added to index
