+++ src/index.js
@@ -1,0 +1,4 @@
+import loadDisplay from './display-component.js';
+import pokemonList from '../data/pokedex.js';
+
+loadDisplay(pokemonList);
