+++ tests/make-card-template.test.js
@@ -1,21 +1,8 @@
+import { makeCardTemplate } from '../src/display-component.js';
 const test = QUnit.test;
 QUnit.module('create pokemon card templates');
 
-function makeCardTemplate(pokemon) {
-   const html = `
-      <li>
-         <h2>${pokemon.pokemon}</h2>
-            <img src="${pokemon.url_image}">
-            <p>
-            <span class="type1">${pokemon.type_1}</span>
-            ${ pokemon.type_2 === 'NA' ? '' : `<span class="type2">${pokemon.type_2}</span>` }
-            </p>
-        </li>
-    `;
-   const template = document.createElement('template');
-   template.innerHTML = html; 
-   return template.content;
-}
+
 
 test('Template test for 2 types', function(assert) {
     // Arrange
